# vector-store/openapi.json
{"openapi":"3.1.0","info":{"title":"Multi-Tenant ChromaDB Server","description":"ChromaDB server with per-project workspaces","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/heartbeat":{"get":{"summary":"Heartbeat","description":"Health check endpoint","operationId":"heartbeat_api_v1_heartbeat_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/version":{"get":{"summary":"Version","description":"Version endpoint","operationId":"version_api_v1_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/projects":{"get":{"summary":"List Projects","description":"List all existing projects","operationId":"list_projects_api_v1_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/{project_name}/collections":{"get":{"summary":"List Collections","description":"List collections for a specific project","operationId":"list_collections_api_v1__project_name__collections_get","parameters":[{"name":"project_name","in":"path","required":true,"schema":{"type":"string","title":"Project Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Collection","description":"Create a new collection in a project","operationId":"create_collection_api_v1__project_name__collections_post","parameters":[{"name":"project_name","in":"path","required":true,"schema":{"type":"string","title":"Project Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{project_name}/collections/{collection_name}":{"get":{"summary":"Get Collection","description":"Get collection information","operationId":"get_collection_api_v1__project_name__collections__collection_name__get","parameters":[{"name":"project_name","in":"path","required":true,"schema":{"type":"string","title":"Project Name"}},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string","title":"Collection Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Collection","description":"Delete a collection","operationId":"delete_collection_api_v1__project_name__collections__collection_name__delete","parameters":[{"name":"project_name","in":"path","required":true,"schema":{"type":"string","title":"Project Name"}},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string","title":"Collection Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{project_name}/collections/{collection_name}/add":{"post":{"summary":"Add To Collection","description":"Add documents to a collection","operationId":"add_to_collection_api_v1__project_name__collections__collection_name__add_post","parameters":[{"name":"project_name","in":"path","required":true,"schema":{"type":"string","title":"Project Name"}},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string","title":"Collection Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{project_name}/collections/{collection_name}/query":{"post":{"summary":"Query Collection","description":"Query a collection","operationId":"query_collection_api_v1__project_name__collections__collection_name__query_post","parameters":[{"name":"project_name","in":"path","required":true,"schema":{"type":"string","title":"Project Name"}},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string","title":"Collection Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/{project_name}/collections/{collection_name}/get":{"get":{"summary":"Get From Collection","description":"Get documents from a collection","operationId":"get_from_collection_api_v1__project_name__collections__collection_name__get_get","parameters":[{"name":"project_name","in":"path","required":true,"schema":{"type":"string","title":"Project Name"}},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string","title":"Collection Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}